requirements.txt
alabaster==0.7.8
apturl==0.5.2
Babel==2.6.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.2
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
feedparser==5.2.1
future==0.18.2
gitsome==0.8.0
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
Jinja2==2.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
selenium==3.141.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
Sphinx==1.8.5
ssh-import-id==5.10
systemd-python==234
torbrowser-launcher==0.3.2
ttystatus==0.38
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webdriver-manager==3.4.2
xkit==0.0.0
xonsh==0.9.13
python-dotenv==0.19.2
